Move PermissionChecker to location, Relocate LocationUpdateCallback to location, chore: Move LocationTracker to location
