+++ app/src/main/java/com/example/capstone_map/location/LocationTracker.java
@@ -1,5 +1,5 @@
 // LocationTracker.java
-package com.example.capstone_map.navigation;
+package com.example.capstone_map.location;
 
 import android.Manifest;
 import android.content.Context;
